improve build performance
- reduce number of api calls
- increase cache duration
