+++ firebase.json
@@ -30,7 +30,7 @@
 				"headers": [
 					{
 						"key": "Cache-Control",
-						"value": "max-age=2592000"
+						"value": "max-age=31536000"
 					}
 				]
 			},
@@ -39,7 +39,7 @@
 				"headers": [
 					{
 						"key": "Cache-Control",
-						"value": "max-age=2592000"
+						"value": "max-age=31536000"
 					}
 				]
 			}
