+++ src/store/index.js
@@ -1,5 +1,4 @@
-import fetch from "isomorphic-unfetch";
-import base64Img from "base64-img";
+import { getAll } from "~/../apiCalls";
 
 export const state = () => ({
 	tags: [],
@@ -44,65 +43,17 @@ export const mutations = {
 	},
 };
 
-function getTags() {
-	return fetch(
-		`https://us-central1-curriculum-vitae-5cd0a.cloudfunctions.net/fastApiProjects/tags`
-	)
-		.then((res) => res.json())
-		.then((data) => data.data);
-}
-
-function getChapters() {
-	return fetch(
-		`https://us-central1-curriculum-vitae-5cd0a.cloudfunctions.net/fastApiProjects/chapters`
-	)
-		.then((res) => res.json())
-		.then((data) => data.data);
-}
-
-async function encodeImage(url) {
-	if (!process.server) {
-		return undefined;
-	}
-
-	return new Promise((resolve, reject) => {
-		base64Img.requestBase64(
-			`https://aridbtumen.cloudimg.io/width/30/x/${url}`,
-			(error, res, body) => {
-				if (error) {
-					return reject(error);
-				}
-				resolve(body);
-			}
-		);
-	});
-}
-
-function getProjects() {
-	return fetch(
-		"https://us-central1-curriculum-vitae-5cd0a.cloudfunctions.net/fastApiProjects/projects"
-	)
-		.then((res) => res.json())
-		.then((data) =>
-			Promise.all(
-				data.data.map(async (project) => {
-					if (project.img) {
-						project.imgPlaceholder = await encodeImage(project.img);
-					}
-					return project;
-				})
-			)
-		);
+function initStore(commit, projects, tags, chapters) {
+	commit("setProjects", projects);
+	commit("setTags", tags);
+	commit("setChapters", chapters);
 }
 
 export const actions = {
-	async nuxtServerInit({ commit }, NuxtContext) {
-		return Promise.all([getProjects(), getTags(), getChapters()]).then(
-			([projects, tags, chapters]) => {
-				commit("setProjects", projects);
-				commit("setTags", tags);
-				commit("setChapters", chapters);
-			}
-		);
+	async nuxtServerInit({ commit }, { payload }) {
+		if (payload) {
+			return initStore(...[commit, ...payload]);
+		}
+		return getAll().then((res) => initStore(...[commit, ...res]));
 	},
 };
